Change argument parser from argparse to click (#2)
* changed CLI tool to click and added entrypoints to setup.cfg

* moved click decorators to main in each module to separate responsibilities

* added extra label and changed to two targets in config and changed log call to be more informative when converting to upload format

* code quality fixes

* code quality darglint

* changed type in click options to click.Path to be able to check if path exists and then convert to pathlib.PAth

* code quality

* removed a mkdir call and added parents flag set to true in downstream mkdir call instead

* moved all click functionality to cli module. removed multiple entrypoints and changed so all calls go through the cli group first

* refactored names of subcommands to be shorter and not clash with imports

* pre-commit checks

* changed wording for copy
