+++ setup.cfg
@@ -68,15 +68,8 @@ testing =
     pytest-cov
 
 [options.entry_points]
-# Add here console scripts like:
-# console_scripts =
-#     script_name = wai_data_tools.module:function
-# For example:
-# console_scripts =
-#     fibonacci = wai_data_tools.skeleton:run
-# And any other entry points, for example:
-# pyscaffold.cli =
-#     awesome = pyscaffoldext.awesome.extension:AwesomeExtension
+console_scripts =
+    wai_data_tools = wai_data_tools.scripts.cli:cli
 
 [tool:pytest]
 # Specify command line options as you would do when invoking pytest directly.
