+++ src/wai_data_tools/scripts/convert_to_upload_format.py
@@ -1,12 +1,9 @@
 """Script for converting file structure to a format that is easier to upload to edge impulse."""
-import argparse
 import logging
 import pathlib
 import shutil
 
 import tqdm
-
-from wai_data_tools.setup_logging import setup_logging
 
 
 def convert_file_structure_to_upload_format(
@@ -16,12 +13,11 @@ def convert_file_structure_to_upload_format(
 
     Args:
         src_root_dir: Source root directory to read files from.
-        dst_root_dir: Destination root directory to store new file
-            structure.
+        dst_root_dir: Destination root directory to store new file structure.
     """
     frame_dirs = [content for content in src_root_dir.iterdir() if content.is_dir()]
 
-    logging.info("Creating new file structure")
+    logging.info("Creating new file structure for uploading to Edge Impulse")
     for frame_dir in tqdm.tqdm(frame_dirs):
         target_dirs = [content for content in frame_dir.iterdir() if content.is_dir()]
         for target_dir in target_dirs:
@@ -34,37 +30,3 @@ def convert_file_structure_to_upload_format(
                 shutil.copy(
                     str(frame_filepath), str(dst_target_dir / frame_filepath.name)
                 )
-
-
-def main():
-    """Entrypoint."""
-    setup_logging()
-
-    parser = argparse.ArgumentParser(
-        "Copies contents of a source file structure and stores it as a format that is "
-        "easier to upload to edge impulse in a destination directory."
-    )
-
-    parser.add_argument(
-        "src_root_dir",
-        type=str,
-        help="Path to the root source directory containing image files",
-    )
-    parser.add_argument(
-        "dst_root_dir",
-        type=str,
-        help="Path to the root destination root directory to save images",
-    )
-
-    args = parser.parse_args()
-
-    src_root_dir = pathlib.Path(args.src_root_dir)
-    dst_root_dir = pathlib.Path(args.dst_root_dir)
-
-    convert_file_structure_to_upload_format(
-        src_root_dir=src_root_dir, dst_root_dir=dst_root_dir
-    )
-
-
-if __name__ == "__main__":
-    main()
